+++ ExpandWorldPrefabs/DelayedSpawn.cs
@@ -4,17 +4,17 @@ using UnityEngine;
 
 namespace ExpandWorld.Prefab;
 
-public class DelayedSpawn(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
+public class DelayedSpawn(float delay, ZdoEntry zdoEntry, bool triggerRules)
 {
   private static readonly List<DelayedSpawn> Spawns = [];
-  public static void Add(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
+  public static void Add(float delay, ZdoEntry zdoEntry, bool triggerRules)
   {
     if (delay <= 0f)
     {
-      Manager.CreateObject(prefab, pos, rot, originalOwner, data, parameters, triggerRules);
+      Manager.CreateObject(zdoEntry, triggerRules);
       return;
     }
-    Spawns.Add(new(delay, pos, rot, prefab, originalOwner, data, parameters, triggerRules));
+    Spawns.Add(new(delay, zdoEntry, triggerRules));
   }
   public static void Execute(float dt)
   {
@@ -32,17 +32,9 @@ public class DelayedSpawn(float delay, Vector3 pos, Quaternion rot, int prefab, 
       Spawns.RemoveAt(i);
     }
   }
-  private readonly Vector3 Pos = pos;
-  private readonly Quaternion Rot = rot;
-  private readonly int Prefab = prefab;
-  private readonly long OriginalOwner = originalOwner;
-  private readonly DataEntry? Data = data;
   public float Delay = delay;
-  private readonly Dictionary<string, string> Parameters = parameters;
-  private readonly bool TriggerRules = triggerRules;
-
   public void Execute()
   {
-    Manager.CreateObject(Prefab, Pos, Rot, OriginalOwner, Data, Parameters, TriggerRules);
+    Manager.CreateObject(zdoEntry, triggerRules);
   }
 }