+++ ExpandWorldPrefabs/service/data/DataHelper.cs
@@ -36,49 +36,6 @@ public class DataHelper
     return false;
   }
   public static DataEntry? Get(string name) => name == "" ? null : DataLoading.Get(name);
-  public static void Init(GameObject obj, Vector3 pos, Quaternion rot, Vector3? scale, DataEntry? data, Dictionary<string, string> pars)
-  {
-    if (data == null && scale == null) return;
-    if (!obj.TryGetComponent<ZNetView>(out var view)) return;
-    var prefab = Utils.GetPrefabName(obj).GetStableHashCode();
-    ZNetView.m_initZDO = ZDOMan.instance.CreateNewZDO(pos, prefab);
-    if (data != null)
-      data.Write([], ZNetView.m_initZDO);
-    ZNetView.m_initZDO.m_rotation = rot.eulerAngles;
-    ZNetView.m_initZDO.Type = view.m_type;
-    ZNetView.m_initZDO.Distant = view.m_distant;
-    ZNetView.m_initZDO.Persistent = view.m_persistent;
-    ZNetView.m_initZDO.m_prefab = prefab;
-    if (view.m_syncInitialScale && scale != null)
-      ZNetView.m_initZDO.Set(ZDOVars.s_scaleHash, scale.Value);
-    ZNetView.m_initZDO.DataRevision = 0;
-    // This is needed to trigger the ZDO sync.
-    ZNetView.m_initZDO.IncreaseDataRevision();
-  }
-  public static ZDO? Init(GameObject obj, Vector3 pos, Quaternion rot, Vector3? scale, DataEntry? data)
-  {
-    if (data == null && scale == null) return null;
-    if (!obj.TryGetComponent<ZNetView>(out var view)) return null;
-    var prefab = Utils.GetPrefabName(obj).GetStableHashCode();
-    ZNetView.m_initZDO = ZDOMan.instance.CreateNewZDO(pos, prefab);
-    if (data != null)
-      data.Write([], ZNetView.m_initZDO);
-    ZNetView.m_initZDO.m_rotation = rot.eulerAngles;
-    ZNetView.m_initZDO.Type = view.m_type;
-    ZNetView.m_initZDO.Distant = view.m_distant;
-    ZNetView.m_initZDO.Persistent = view.m_persistent;
-    ZNetView.m_initZDO.m_prefab = prefab;
-    if (view.m_syncInitialScale && scale != null)
-      ZNetView.m_initZDO.Set(ZDOVars.s_scaleHash, scale.Value);
-    ZNetView.m_initZDO.DataRevision = 0;
-    // This is needed to trigger the ZDO sync.
-    ZNetView.m_initZDO.IncreaseDataRevision();
-    return ZNetView.m_initZDO;
-  }
-  public static void CleanUp()
-  {
-    ZNetView.m_initZDO = null;
-  }
 
   public static List<string>? GetValuesFromGroup(string group)
   {